move unit test to exising file
